% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cggr.R
\name{cggr}
\alias{cggr}
\title{Run canonical Gaussian graphical regression with covariates}
\usage{
cggr(
  responses,
  covariates,
  asparse,
  nregmean = 10,
  nlambda = 100,
  lambdafactor = 1e-04,
  maxit = 3e+06,
  tol = 1e-08,
  nfolds = 5,
  verbose = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{responses}{\eqn{n x p} matrix of responses}

\item{covariates}{\eqn{n x q} matrix of covariates}

\item{asparse}{The sparse-group lasso mixing parameter with \eqn{0\leq \alpha \leq 1}}

\item{nregmean}{The number of eta values - default is 10}

\item{nlambda}{The number of lambda values - default is 100}

\item{lambdafactor}{The smallest value of lambda as a fraction of the maximum lambda.}

\item{maxit}{The maximum number of iterations. Default is \eqn{3\times 10^6}}

\item{tol}{The convergence threshhold for optimization. Default is \eqn{10^{-8}}}

\item{nfolds}{Number of folds for cross-validation. Default is 5.}

\item{verbose}{If TRUE, prints progress messages. Default is FALSE.}

\item{parallel}{If TRUE, runs the nodewise regressions in parallel. Default is TRUE.}
}
\description{
Run canonical Gaussian graphical regression with covariates
}
