% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncagr-methods.R
\name{coef.ncagr}
\alias{coef.ncagr}
\title{Extract estimated coefficients from a 'ncagr' object}
\usage{
\method{coef}{ncagr}(fit)
}
\arguments{
\item{fit}{An object of class [ncagr].}
}
\value{
A list with components 'beta', an array of precision matrix
 components, and 'gamma', the estimate mean component.
}
\description{
Extract estimated coefficients from a 'ncagr' object
}
