% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncagr.R
\name{ncagr}
\alias{ncagr}
\title{Run Natural Covariate-adjusted Graphical Regression}
\usage{
ncagr(
  responses,
  covariates,
  gmixpath = seq(0, 1, by = 0.1),
  sglmixpath = 0.75,
  nlambda = 100,
  lambdafactor = 1e-08,
  maxit = 3e+06,
  tol = 1e-06,
  nfolds = 5,
  verbose = TRUE,
  ncores = 1,
  adaptive = FALSE
)
}
\arguments{
\item{responses}{\eqn{n \times p} matrix of responses}

\item{covariates}{\eqn{n \times q} matrix of covariates}

\item{gmixpath}{A path of ncagr mixing parameters. Default is 0.1, 0.2, ..., 0.9.}

\item{sglmixpath}{A path of sparse-group lasso mixing parameter with \eqn{0\leq \alpha \leq 1}. Default is 0.75.}

\item{nlambda}{The number of lambda values to use for cross-validation - default is 100.}

\item{lambdafactor}{The smallest value of lambda as a fraction of the maximum lambda.}

\item{maxit}{The maximum number of iterations. Default is \eqn{3\times 10^6}.}

\item{tol}{The convergence threshhold for optimization. Default is \eqn{10^{-6}}.}

\item{nfolds}{Number of folds for cross-validation. Default is 5.}

\item{verbose}{If TRUE, prints progress messages. Default is TRUE.}

\item{ncores}{Runs the nodewise regressions in parallel using that many cores. Default is 1.}

\item{adaptive}{Use adaptive weights when fitting nodewise regressions. Default is FALSE.}
}
\description{
Run Natural Covariate-adjusted Graphical Regression
}
