% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncagr-methods.R
\name{predict.ncagr}
\alias{predict.ncagr}
\title{Predict from a 'ncagr' object}
\usage{
\method{predict}{ncagr}(fit, u)
}
\arguments{
\item{fit}{An object of class [ncagr].}

\item{u}{A vector of covariates.}
}
\value{
A list with components 'precision', the predicted precision matrix,
 and 'mean', the predicted mean given the new covariates.
}
\description{
This function predicts the mean and precision matrix for a new vector
 of covariates.
}
